Fix validation of provided id in all routes and fix the const naming conflict in seats put method

+++ routes/concerts.routes.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const Concert = require('../models/concert.model');
+const mongoose = require('mongoose');
 
 router.get('/concerts', async (req, res) => {
   try {
@@ -13,9 +14,13 @@ router.get('/concerts', async (req, res) => {
 
 router.get('/concerts/:id', async (req, res) => {
   try {
-    const concert = await Concert.findById(req.params.id);
-    if (!concert) res.status(404).json({ message: 'Not found' });
-    else res.json(concert);
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      res.status(404).json({ message: 'Provided id is not valid' });
+    } else {
+      const concert = await Concert.findById(req.params.id);
+      if (!concert) res.status(404).json({ message: 'Not found' });
+      else res.json(concert);
+    }
   }
   catch (err) {
     res.status(500).json({ message: err });
@@ -39,37 +44,45 @@ router.post('/concerts', async (req, res) => {
   }
 });
 
-router.delete('/concerts/:id', async (req, res) => {
+router.put('/concerts/:id', async (req, res) => {
+  const { performer, genre, price, day, image } = req.body;
   try {
-    const concert = await Concert.findById(req.params.id);
-    if (concert) {
-      await Concert.deleteOne({ _id: req.params.id });
-      res.json({ message: 'OK' });
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      res.status(404).json({ message: 'Provided id is not valid' });
+    } else {
+      const concert = await Concert.findById(req.params.id);
+      if (concert) {
+        await Concert.updateOne({ _id: req.params.id }, {
+          $set: {
+            performer: performer,
+            genre: genre,
+            price: price,
+            day: day,
+            image: image,
+          }
+        });
+        res.json({ message: 'OK' });
+      }
+      else res.status(404).json({ message: 'Not found...' });
     }
-    else res.status(404).json({ message: 'Not found...' });
   }
   catch (err) {
     res.status(500).json({ message: err });
   }
 });
 
-router.put('/concerts/:id', async (req, res) => {
-  const { performer, genre, price, day, image } = req.body;
+router.delete('/concerts/:id', async (req, res) => {
   try {
-    const concert = await Concert.findById(req.params.id);
-    if (concert) {
-      await Concert.updateOne({ _id: req.params.id }, {
-        $set: {
-          performer: performer,
-          genre: genre,
-          price: price,
-          day: day,
-          image: image,
-        }
-      });
-      res.json({ message: 'OK' });
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      res.status(404).json({ message: 'Provided id is not valid' });
+    } else {
+      const concert = await Concert.findById(req.params.id);
+      if (concert) {
+        await Concert.deleteOne({ _id: req.params.id });
+        res.json({ message: 'OK' });
+      }
+      else res.status(404).json({ message: 'Not found...' });
     }
-    else res.status(404).json({ message: 'Not found...' });
   }
   catch (err) {
     res.status(500).json({ message: err });
