+++ routes/testimonials.routes.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const Testimonial = require('../models/testimonial.model');
+const mongoose = require('mongoose');
 
 router.get('/testimonials', async (req, res) => {
   try {
@@ -26,9 +27,13 @@ router.get('/testimonials/random', async (req, res) => {
 
 router.get('/testimonials/:id', async (req, res) => {
   try {
-    const testimonial = await Testimonial.findById(req.params.id);
-    if (!testimonial) res.status(404).json({ message: 'Not found' });
-    else res.json(testimonial);
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      res.status(404).json({ message: 'Provided id is not valid' });
+    } else {
+      const testimonial = await Testimonial.findById(req.params.id);
+      if (!testimonial) res.status(404).json({ message: 'Not found' });
+      else res.json(testimonial);
+    }
   }
   catch (err) {
     res.status(500).json({ message: err });
@@ -52,17 +57,21 @@ router.post('/testimonials', async (req, res) => {
 router.put('/testimonials/:id', async (req, res) => {
   const { author, text } = req.body;
   try {
-    const testimonial = await Testimonial.findById(req.params.id);
-    if (testimonial) {
-      await Testimonial.updateOne({ _id: req.params.id }, {
-        $set: {
-          author: author,
-          text: text,
-        }
-      });
-      res.json({ message: 'OK' });
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      res.status(404).json({ message: 'Provided id is not valid' });
+    } else {
+      const testimonial = await Testimonial.findById(req.params.id);
+      if (testimonial) {
+        await Testimonial.updateOne({ _id: req.params.id }, {
+          $set: {
+            author: author,
+            text: text,
+          }
+        });
+        res.json({ message: 'OK' });
+      }
+      else res.status(404).json({ message: 'Not found...' });
     }
-    else res.status(404).json({ message: 'Not found...' });
   }
   catch (err) {
     res.status(500).json({ message: err });
@@ -71,12 +80,16 @@ router.put('/testimonials/:id', async (req, res) => {
 
 router.delete('/testimonials/:id', async (req, res) => {
   try {
-    const testimonial = await Testimonial.findById(req.params.id);
-    if (testimonial) {
-      await Testimonial.deleteOne({ _id: req.params.id });
-      res.json({ message: 'OK' });
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      res.status(404).json({ message: 'Provided id is not valid' });
+    } else {
+      const testimonial = await Testimonial.findById(req.params.id);
+      if (testimonial) {
+        await Testimonial.deleteOne({ _id: req.params.id });
+        res.json({ message: 'OK' });
+      }
+      else res.status(404).json({ message: 'Not found...' });
     }
-    else res.status(404).json({ message: 'Not found...' });
   }
   catch (err) {
     res.status(500).json({ message: err });
